requirements.txt
openai-whisper==20230124
yt-dlp 
pyinstaller 

srt
deep_translator
janome